feat(state): complete Svelte 5 class-based state adapters and enforce UI/persisted boundary (#451)
* Initial plan

* feat(state): complete Svelte 5 class-based adapters and decouple UI/persisted state (#231 #232 #233)

Agent-Logs-Url: https://github.com/nextxm/paisa/sessions/b92a5a4d-139f-4317-bcf7-4c6ed5581cd3


* fix(lint): format yoy page with prettier to fix CI lint failure

Agent-Logs-Url: https://github.com/nextxm/paisa/sessions/0446d8b1-7fb1-4fd1-8ae9-e4bc1654e456


---------
