+++ src/store.ts
@@ -1,5 +1,6 @@
 import { writable, derived, get } from "svelte/store";
 import * as d3 from "d3";
+import { cashflowExpenseDepth, cashflowIncomeDepth } from "./persisted_store";
 
 import dayjs from "dayjs";
 import type { AccountTfIdf, LedgerFileError, SheetFileError, SheetLineResult } from "$lib/utils";
@@ -142,6 +143,23 @@ export function setAllowedDateRange(dates: dayjs.Dayjs[]) {
   if (start) {
     dateMin.set(start);
     dateMax.set(end);
+  }
+}
+
+// Transient UI state for cashflow depth controls (not persisted).
+// The allowed range is set by the page after loading data; the selected depth is persisted.
+export const cashflowExpenseDepthAllowed = writable({ min: 1, max: 1 });
+export const cashflowIncomeDepthAllowed = writable({ min: 1, max: 1 });
+
+export function setCashflowDepthAllowed(expense: number, income: number) {
+  cashflowExpenseDepthAllowed.set({ min: 1, max: expense });
+  if (get(cashflowExpenseDepth) == 0 || get(cashflowExpenseDepth) > expense) {
+    cashflowExpenseDepth.set(expense);
+  }
+
+  cashflowIncomeDepthAllowed.set({ min: 1, max: income });
+  if (get(cashflowIncomeDepth) == 0 || get(cashflowIncomeDepth) > income) {
+    cashflowIncomeDepth.set(income);
   }
 }
